remove product origin converter
