+++ server/src/main/java/org/elasticsearch/common/logging/TraceIdConverter.java
@@ -8,17 +8,21 @@
 
 package org.elasticsearch.common.logging;
 
+import org.apache.logging.log4j.core.LogEvent;
 import org.apache.logging.log4j.core.config.plugins.Plugin;
 import org.apache.logging.log4j.core.pattern.ConverterKeys;
+import org.apache.logging.log4j.core.pattern.LogEventPatternConverter;
 import org.apache.logging.log4j.core.pattern.PatternConverter;
 import org.elasticsearch.tasks.Task;
+
+import java.util.Objects;
 
 /**
  * Pattern converter to format the trace id provided in the traceparent header into JSON fields <code>trace.id</code>.
  */
 @Plugin(category = PatternConverter.CATEGORY, name = "TraceIdConverter")
 @ConverterKeys({ "trace_id" })
-public final class TraceIdConverter extends ThreadContextBasedConverter {
+public final class TraceIdConverter extends LogEventPatternConverter {
     /**
      * Called by log4j2 to initialize this converter.
      */
@@ -27,6 +31,29 @@ public final class TraceIdConverter extends ThreadContextBasedConverter {
     }
 
     public TraceIdConverter() {
-        super("trace_id", Task.TRACE_ID);
+        super("trace_id", "trace_id");
     }
+
+    public static String getTraceId() {
+        return HeaderWarning.THREAD_CONTEXT.stream()
+            .map(t -> t.<String>getHeader(Task.TRACE_ID))
+            .filter(Objects::nonNull)
+            .findFirst()
+            .orElse(null);
+    }
+
+    /**
+     * Formats the trace.id into json fields.
+     *
+     * @param event - a log event is ignored in this method as it uses the clusterId value
+     *              from <code>NodeAndClusterIdStateListener</code> to format
+     */
+    @Override
+    public void format(LogEvent event, StringBuilder toAppendTo) {
+        String traceId = getTraceId();
+        if (traceId != null) {
+            toAppendTo.append(traceId);
+        }
+    }
+
 }
